v0.0.3 (3/6/2015)
Update documentation
Include .gitattributes
Extract iteration methods to DTOIterator class
