+++ .gitattributes
@@ -1,0 +1,6 @@
+/.gitattributes     export-ignore
+/.gitignore         export-ignore
+/.travis.yml        export-ignore
+/docs               export-ignore
+/phpunit.xml        export-ignore
+/tests              export-ignore
